// plataforma-promotores/frontend/src/components/common/EstadoBadge.d.ts
import React from 'react';
import { TramiteEstado, TramiteEstadoVista } from '../../types';
interface EstadoBadgeProps {
    estado: TramiteEstado | TramiteEstadoVista;
    size?: 'sm' | 'md';
}
export declare const EstadoBadge: React.FC<EstadoBadgeProps>;
export {};
//# sourceMappingURL=EstadoBadge.d.ts.map